[kafkaadmin] moved topic states
